requirements.txt
openai
python-dotenv
PyGithub
gitpython
